Use the layout from the old branch
