+++ crates/components/src/map_view.rs
@@ -79,7 +79,7 @@ impl MapView {
         let tileset = &tilesets.data[map.tileset_id];
 
         let mut passages = luminol_data::Table2::new(map.data.xsize(), map.data.ysize());
-        luminol_graphics::collision::calculate_passages(
+        luminol_graphics::primitives::collision::calculate_passages(
             &tileset.passages,
             &tileset.priorities,
             &map.data,
