+++ crates/graphics/src/loaders/atlas.rs
@@ -14,7 +14,7 @@
 //
 // You should have received a copy of the GNU General Public License
 // along with Luminol.  If not, see <http://www.gnu.org/licenses/>.
-use crate::{tiles::Atlas, GraphicsState};
+use crate::{Atlas, GraphicsState};
 
 #[derive(Default)]
 pub struct Loader {
