+++ crates/graphics/src/primitives/collision/collision.wgsl
