+++ crates/graphics/src/primitives/collision/vertex.rs
