+++ crates/graphics/src/primitives/grid/display.rs
