+++ crates/graphics/src/primitives/sprite/graphic.rs
