+++ crates/graphics/src/primitives/sprite/shader.rs
@@ -15,33 +15,30 @@
 // You should have received a copy of the GNU General Public License
 // along with Luminol.  If not, see <http://www.gnu.org/licenses/>.
 
-use crate::{vertex::Vertex, BindGroupLayouts};
+use std::collections::HashMap;
+
+use naga_oil::compose::ComposerError;
+
+use crate::{primitives::BindGroupLayouts, Vertex};
 
 fn create_shader(
+    composer: &mut naga_oil::compose::Composer,
     render_state: &luminol_egui_wgpu::RenderState,
     bind_group_layouts: &BindGroupLayouts,
     target: wgpu::BlendState,
-) -> wgpu::RenderPipeline {
+) -> Result<wgpu::RenderPipeline, ComposerError> {
     let push_constants_supported = crate::push_constants_supported(render_state);
 
-    let mut composer = naga_oil::compose::Composer::default().with_capabilities(
-        push_constants_supported
-            .then_some(naga::valid::Capabilities::PUSH_CONSTANT)
-            .unwrap_or_default(),
-    );
-
-    let module = composer
-        .make_naga_module(naga_oil::compose::NagaModuleDescriptor {
-            source: include_str!("sprite.wgsl"),
-            file_path: "sprite.wgsl",
-            shader_type: naga_oil::compose::ShaderType::Wgsl,
-            shader_defs: std::collections::HashMap::from([(
-                "USE_PUSH_CONSTANTS".to_string(),
-                naga_oil::compose::ShaderDefValue::Bool(push_constants_supported),
-            )]),
-            additional_imports: &[],
-        })
-        .expect("failed to create sprite shader module");
+    let module = composer.make_naga_module(naga_oil::compose::NagaModuleDescriptor {
+        source: include_str!("sprite.wgsl"),
+        file_path: "sprite.wgsl",
+        shader_type: naga_oil::compose::ShaderType::Wgsl,
+        shader_defs: HashMap::from([(
+            "USE_PUSH_CONSTANTS".to_string(),
+            naga_oil::compose::ShaderDefValue::Bool(push_constants_supported),
+        )]),
+        additional_imports: &[],
+    })?;
 
     let shader_module = render_state
         .device
@@ -80,7 +77,7 @@ fn create_shader(
                 push_constant_ranges,
             });
 
-    render_state
+    Ok(render_state
         .device
         .create_render_pipeline(&wgpu::RenderPipelineDescriptor {
             label: Some("Tilemap Sprite Render Pipeline"),
@@ -105,55 +102,51 @@ fn create_shader(
             depth_stencil: None,
             multisample: wgpu::MultisampleState::default(),
             multiview: None,
-        })
+        }))
 }
 
+const BLEND_ADD: wgpu::BlendState = wgpu::BlendState {
+    color: wgpu::BlendComponent {
+        src_factor: wgpu::BlendFactor::SrcAlpha,
+        dst_factor: wgpu::BlendFactor::One,
+        operation: wgpu::BlendOperation::Add,
+    },
+    alpha: wgpu::BlendComponent {
+        src_factor: wgpu::BlendFactor::One,
+        dst_factor: wgpu::BlendFactor::One,
+        operation: wgpu::BlendOperation::Add,
+    },
+};
+const BLEND_SUBTRACT: wgpu::BlendState = wgpu::BlendState {
+    color: wgpu::BlendComponent {
+        src_factor: wgpu::BlendFactor::SrcAlpha,
+        dst_factor: wgpu::BlendFactor::One,
+        operation: wgpu::BlendOperation::ReverseSubtract,
+    },
+    alpha: wgpu::BlendComponent {
+        src_factor: wgpu::BlendFactor::Zero,
+        dst_factor: wgpu::BlendFactor::One,
+        operation: wgpu::BlendOperation::ReverseSubtract,
+    },
+};
+
 pub fn create_sprite_shaders(
+    composer: &mut naga_oil::compose::Composer,
     render_state: &luminol_egui_wgpu::RenderState,
     bind_group_layouts: &BindGroupLayouts,
-) -> std::collections::HashMap<luminol_data::BlendMode, wgpu::RenderPipeline> {
+) -> Result<HashMap<luminol_data::BlendMode, wgpu::RenderPipeline>, ComposerError> {
     [
         (
             luminol_data::BlendMode::Normal,
             wgpu::BlendState::ALPHA_BLENDING,
         ),
-        (
-            luminol_data::BlendMode::Add,
-            wgpu::BlendState {
-                color: wgpu::BlendComponent {
-                    src_factor: wgpu::BlendFactor::SrcAlpha,
-                    dst_factor: wgpu::BlendFactor::One,
-                    operation: wgpu::BlendOperation::Add,
-                },
-                alpha: wgpu::BlendComponent {
-                    src_factor: wgpu::BlendFactor::One,
-                    dst_factor: wgpu::BlendFactor::One,
-                    operation: wgpu::BlendOperation::Add,
-                },
-            },
-        ),
-        (
-            luminol_data::BlendMode::Subtract,
-            wgpu::BlendState {
-                color: wgpu::BlendComponent {
-                    src_factor: wgpu::BlendFactor::SrcAlpha,
-                    dst_factor: wgpu::BlendFactor::One,
-                    operation: wgpu::BlendOperation::ReverseSubtract,
-                },
-                alpha: wgpu::BlendComponent {
-                    src_factor: wgpu::BlendFactor::Zero,
-                    dst_factor: wgpu::BlendFactor::One,
-                    operation: wgpu::BlendOperation::ReverseSubtract,
-                },
-            },
-        ),
+        (luminol_data::BlendMode::Add, BLEND_ADD),
+        (luminol_data::BlendMode::Subtract, BLEND_SUBTRACT),
     ]
     .into_iter()
     .map(|(mode, target)| {
-        (
-            mode,
-            create_shader(render_state, bind_group_layouts, target),
-        )
+        let shader = create_shader(composer, render_state, bind_group_layouts, target)?;
+        Ok((mode, shader))
     })
     .collect()
 }
