+++ crates/graphics/src/primitives/sprite/sprite.wgsl
