+++ crates/graphics/src/primitives/sprite/vertices.rs
@@ -15,7 +15,7 @@
 // You should have received a copy of the GNU General Public License
 // along with Luminol.  If not, see <http://www.gnu.org/licenses/>.
 
-use crate::quad::Quad;
+use crate::Quad;
 
 #[derive(Debug)]
 pub struct Vertices {
