+++ crates/graphics/src/primitives/tiles/autotile_ids.rs
@@ -1,0 +1,70 @@
+// Copyright (C) 2023 Lily Lyons
+//
+// This file is part of Luminol.
+//
+// Luminol is free software: you can redistribute it and/or modify
+// it under the terms of the GNU General Public License as published by
+// the Free Software Foundation, either version 3 of the License, or
+// (at your option) any later version.
+//
+// Luminol is distributed in the hope that it will be useful,
+// but WITHOUT ANY WARRANTY; without even the implied warranty of
+// MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+// GNU General Public License for more details.
+//
+// You should have received a copy of the GNU General Public License
+// along with Luminol.  If not, see <http://www.gnu.org/licenses/>.
+
+/// Hardcoded list of tiles from r48 and old python Luminol.
+/// There seems to be very little pattern in autotile IDs so this is sadly
+/// the best we can do.
+pub const AUTOTILES: [[u32; 4]; 48] = [
+    [26, 27, 32, 33],
+    [4, 27, 32, 33],
+    [26, 5, 32, 33],
+    [4, 5, 32, 33],
+    [26, 27, 32, 11],
+    [4, 27, 32, 11],
+    [26, 5, 32, 11],
+    [4, 5, 32, 11],
+    [26, 27, 10, 33],
+    [4, 27, 10, 33],
+    [26, 5, 10, 33],
+    [4, 5, 10, 33],
+    [26, 27, 10, 11],
+    [4, 27, 10, 11],
+    [26, 5, 10, 11],
+    [4, 5, 10, 11],
+    [24, 25, 30, 31],
+    [24, 5, 30, 31],
+    [24, 25, 30, 11],
+    [24, 5, 30, 11],
+    [14, 15, 20, 21],
+    [14, 15, 20, 11],
+    [14, 15, 10, 21],
+    [14, 15, 10, 11],
+    [28, 29, 34, 35],
+    [28, 29, 10, 35],
+    [4, 29, 34, 35],
+    [4, 29, 10, 35],
+    [38, 39, 44, 45],
+    [4, 39, 44, 45],
+    [38, 5, 44, 45],
+    [4, 5, 44, 45],
+    [24, 29, 30, 35],
+    [14, 15, 44, 45],
+    [12, 13, 18, 19],
+    [12, 13, 18, 11],
+    [16, 17, 22, 23],
+    [16, 17, 10, 23],
+    [40, 41, 46, 47],
+    [4, 41, 46, 47],
+    [36, 37, 42, 43],
+    [36, 5, 42, 43],
+    [12, 17, 18, 23],
+    [12, 13, 42, 43],
+    [36, 41, 42, 47],
+    [16, 17, 46, 47],
+    [12, 17, 42, 47],
+    [0, 1, 6, 7],
+];
